Moved mock classes to separated files
