+++ test/unit/mock/blockchain-module-manager-mock.js
@@ -1,0 +1,17 @@
+import { ethers } from 'ethers';
+
+class BlockchainModuleManagerMock {
+    getR2() {
+        return 20;
+    }
+
+    getR0() {
+        return 3;
+    }
+
+    convertToWei(blockchainId, value) {
+        return ethers.utils.parseUnits(value.toString(), 'ether').toString();
+    }
+}
+
+export default BlockchainModuleManagerMock;
