+++ test/unit/mock/event-emitter-mock.js
@@ -1,0 +1,3 @@
+class EventEmitterMock {}
+
+export default EventEmitterMock;
