+++ test/unit/mock/network-module-manager-mock.js
@@ -1,0 +1,3 @@
+class NetworkModuleManagerMock {}
+
+export default NetworkModuleManagerMock;
