+++ test/unit/mock/validation-module-manager-mock.js
@@ -1,0 +1,14 @@
+import { ethers } from 'ethers';
+
+class ValidationModuleManagerMock {
+    callHashFunction(data) {
+        const bytesLikeData = ethers.utils.toUtf8Bytes(data);
+        return ethers.utils.sha256(bytesLikeData);
+    }
+
+    getHashFunctionName() {
+        return 'sha256';
+    }
+}
+
+export default ValidationModuleManagerMock;
